FastAPI_Main/pyproject.toml
[tool.poetry]
name = "fastapi-main"
version = "0.1.0"
description = "Migrated from Django API CRUD"
authors = ["Akshat"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
daphne = "^4.1.2"
wfastcgi = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

